final commit

+++ Library/PublishersForms/PublisherListForm.cs
@@ -1,12 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Security.Policy;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows.Forms;
 using Library.Domain.Repositories;
 using Publisher = Library.Data.Entities.Models.Publisher;
@@ -22,7 +14,6 @@ namespace Library.PublishersForms
             InitializeComponent();
             PublisherRepository = publisherRepository;
             BookRepository = bookRepository;
-            AddRefreshList();
         }
 
         public void AddRefreshList()
@@ -34,34 +25,57 @@ namespace Library.PublishersForms
         private void AddPublisherButtonClick(object sender, EventArgs e)
         {
             var publisherCreateForm = new PublisherCreateForm(PublisherRepository);
-            publisherCreateForm.Show();
+            publisherCreateForm.ShowDialog();
             AddRefreshList();
         }
 
         private void EditPublisherButtonClick(object sender, EventArgs e)
         {
-            var selectedPublisher = publishersListBox.SelectedItem as Publisher;
-            var publisherEditForm = new PublisherEditForm(PublisherRepository){SelectedPublisher = selectedPublisher};
-            publisherEditForm.SetText();
-            publisherEditForm.ShowDialog();
-            AddRefreshList();
+            if (!(publishersListBox.SelectedItem is Publisher selectedPublisher))
+            {
+                var errorForm = new ErrorForm("You need to select publisher");
+                errorForm.ShowDialog();
+            }
+            else
+            {
+                var publisherEditForm = new PublisherEditForm(PublisherRepository)
+                    {SelectedPublisher = selectedPublisher};
+                publisherEditForm.SetText();
+                publisherEditForm.ShowDialog();
+                AddRefreshList();
+            }
         }
 
         private void PublishersDeleteButtonClick(object sender, EventArgs e)
         {
-            var selectedPublisher = publishersListBox.SelectedItem as Publisher;
-            PublisherRepository.RemovePublisher(selectedPublisher.PublisherId);
-            AddRefreshList();
+            if (!(publishersListBox.SelectedItem is Publisher selectedPublisher))
+            {
+                var errorForm = new ErrorForm("You need to select publisher");
+                errorForm.ShowDialog();
+            }
+            else
+            {
+                PublisherRepository.RemovePublisher(selectedPublisher.PublisherId);
+                AddRefreshList();
+            }
         }
 
         private void PublishersDetailsButtonClick(object sender, EventArgs e)
         {
-            var selectedPublisher = publishersListBox.SelectedItem as Publisher;
-            var publisherDetailsForm = new PublisherDetailsForm(PublisherRepository,BookRepository) { SelectedPublisher = selectedPublisher };
-            publisherDetailsForm.AddRefreshList();
-            publisherDetailsForm.ShowDialog();
-            AddRefreshList();
-            
+            if (!(publishersListBox.SelectedItem is Publisher selectedPublisher))
+            {
+                var errorForm = new ErrorForm("You need to select publisher");
+                errorForm.ShowDialog();
+            }
+            else
+            {
+                var publisherDetailsForm = new PublisherDetailsForm(PublisherRepository, BookRepository)
+                    {SelectedPublisher = selectedPublisher};
+                publisherDetailsForm.AddRefreshList();
+                publisherDetailsForm.ShowDialog();
+                AddRefreshList();
+            }
+
         }
     }
 }
